Separate test and modify docs
