+++ docs/source/modules/index.rst
@@ -4,7 +4,11 @@ References
 .. toctree::
    :maxdepth: 3
 
-.. automodule:: iiblib.iibclient
+.. automodule:: iiblib.iib_client
+.. automodule:: iiblib.iib_authentication
+.. automodule:: iiblib.iib_build_details_pager
+.. automodule:: iiblib.iib_build_details_model
+.. automodule:: iiblib.iib_session
    :members:
    :show-inheritance:
    :inherited-members:
